perf(webapp): componetization of expense table view
