+++ webapp/src/routes/ExpenseReport/index.js
@@ -5,8 +5,7 @@ import {
   FormControlLabel,
   RadioGroup,
   Radio,
-  Divider,
-  Typography
+  Divider
 } from '@mui/material'
 import { useTranslation } from 'react-i18next'
 
@@ -14,14 +13,12 @@ import useExpenseReportState from '../../hooks/customHooks/useExpenseReportState
 import TreasuryBalance from '../../components/TreasuryBalance'
 import BarChartReport from '../../components/BarChartReport'
 import PieChartReport from '../../components/PieChartReport'
-import { formatWithThousandSeparator } from '../../utils'
-import TableReport from '../../components/TableReport'
 import SelectComponent from '../../components/Select'
 
 import styles from './styles'
+import ExpenseTableReport from './expense-table-report'
 
 const useStyles = makeStyles(styles)
-const rowsCenter = { flex: 1, align: 'center', headerAlign: 'center' }
 
 const ExpenseReport = () => {
   const classes = useStyles()
@@ -43,59 +40,6 @@ const ExpenseReport = () => {
     showElectionRadio === 'allElections'
       ? expenseByElectionsList
       : delegatesList
-
-  const columns = [
-    {
-      field: 'election',
-      hide: !tableData[0]?.election,
-      headerName: t('tableElectionHeader', { ns: 'expenseRoute' }),
-      cellClassName: classes.links,
-      ...rowsCenter
-    },
-    {
-      field: 'name',
-      hide: !tableData[0]?.color,
-      headerName: tableData[0]?.name
-        ? t('tableHeader1', { ns: 'expenseRoute' })
-        : t('tableElectionHeader', { ns: 'expenseRoute' }),
-      cellClassName: classes.links,
-      renderCell: param => (
-        <a
-          className={tableData[0]?.name ? '' : classes.disableLink}
-          href={`https://eosdetective.io/network/transfers?accounts=${param.value}&time_min=1661975129190&time_max=1669751129190&excludedAccounts=&excludedCategories=system`}
-        >
-          {param.value}
-        </a>
-      ),
-      ...rowsCenter
-    },
-    {
-      field: tableData[0]?.election ? 'EOS_TOTAL' : 'EOS_CATEGORIZED',
-      headerName: 'EOS',
-      renderCell: param => <>{formatWithThousandSeparator(param.value, 2)}</>,
-      type: 'number',
-      ...rowsCenter
-    },
-    {
-      field: tableData[0]?.election ? 'USD_TOTAL' : 'USD_CATEGORIZED',
-      headerName: 'USD',
-      renderCell: param => <>{formatWithThousandSeparator(param.value, 2)}</>,
-      type: 'number',
-      ...rowsCenter
-    },
-    {
-      field: 'EOS_CATEGORIZED_PERCENT',
-      headerName: t('tableHeader7', { ns: 'expenseRoute' }),
-      type: 'number',
-      ...rowsCenter
-    },
-    {
-      field: 'EOS_UNCATEGORIZED_PERCENT',
-      headerName: t('tableHeader8', { ns: 'expenseRoute' }),
-      type: 'number',
-      ...rowsCenter
-    }
-  ]
 
   return (
     <div className={classes.root}>
@@ -164,18 +108,10 @@ const ExpenseReport = () => {
         </div>
       </div>
       <Divider variant="middle" />
-      <div className={classes.tableContainer}>
-        <div className={classes.title}>
-          <Typography variant="span">
-            {showElectionRadio === 'allElections'
-              ? t('titleTable2', { ns: 'expenseRoute' })
-              : t('titleTable', { ns: 'expenseRoute' })}
-          </Typography>
-          <div id="id-table-container">
-            <TableReport columns={columns} dataPercent={tableData} />
-          </div>
-        </div>
-      </div>
+      <ExpenseTableReport
+        tableData={tableData}
+        showElectionRadio={showElectionRadio}
+      />
     </div>
   )
 }
